+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Next Release
 
+#### Fixes
+
+- fix: Make sure `LogEntry.changes_dict()` returns an empty dict instead of `None` when `json.loads()` returns `None`. ([#472](https://github.com/jazzband/django-auditlog/pull/472))
+- feat: Make timestamp in LogEntry overwritable. ([#476](https://github.com/jazzband/django-auditlog/pull/476))
+
 ## 2.2.2 (2023-01-16)
 
 #### Fixes